cmd: initial support subcmd (#814)
Commands are now represented by a global map. When the linter is launched, the standard NoVerify commands are registered in it. Further, if the OverriddenCommands field in the config is not empty, then commands are taken from it and redefined in the global map.
This will make it easy to add and override commands from NoVerify-based tools.