linter: part of the analysis and indexing of the class is divided into different functions
So, there are now two functions handleClass and checkClass.
Each of them, ideally (in the future), can only call checking
and indexing functions, respectively. This will allow us to
further separate the stages of indexing and verification more
easily.

The new root_walker_checkers file was created in order to start
storing everything related only to checks in the RootWalker in it,
again for easier separation in the future.

The getClass function has been renamed to getOrCreateCurrentClass
to better match the function's purpose.
Also, now if a function does not find a class in the meta of the
current file, then it looks in the general meta.
This is necessary, because earlier, even during analysis, when
traversing a class, the class was added to the meta of the current
file, while it was already in the global meta. Now the search for
a class at the analysis stage is also performed in the global meta.