meta,linter: use TypesMap.Map instead of TypesMap.AppendString
The TypesMap.AppendString may mutate the map, which makes it harder
for us to migrate to fully immutable type maps.

The new Map() method is supposed to be used in places where we
want to create a new types map based on other types map.

All current AppendString usages were replaced and the method itself
is removed now.