linter: small refactor
Separated indexing and parsing logic for TraitUseStmt.

The RootWalker.getOrCreateCurrentClass method has been
renamed to getCurrentClass and now does not create a
class if it is not found. Since earlier the logic of
adding classes was built on this, now it is redundant.

Temporarily disabled changing the types of class properties
when analyzing the assignment to them, since this changes
the class in the global meta and leads to data races.