cmd: small refactor rule init code (#916)
The Config.Rules field is now fully initialized, including
the fields of the rules.Set structure. Thus, we got rid of the
need to initialize this field every time.

The InitEmbeddedRules function has been renamed to AddEmbeddedRules
and now accepts not the config, but the rules directly, as well
as for the appendRuleSet function.