linter: refactor
Functions for generating names for shape and
closure have been moved to the autogen package.
In the future, it is planned to add the generation
of anonymous class names to this package.

For shape, names are now generated based on the
keys they are composed of. This will decouple
the generation of names from the context.

Before:
\shape$filename.php$count_shapes$

Now:
\shape$x:int,y:string$