linter,meta: move all types-related code to `types` package
There are a lot of changes since all `meta.TypesMap` things were
renamed to `types.Map` (and there more examples like this).

One benefit is that if we decouple types and meta it
will be easier to further decompose indexdata types and
meta types (they both would need to use `types`, so we
more it to a separate package to avoid circular deps).