cmd: parsing of flags now happens at the beginning, and parsing of rules happens after
Now the help in flags does not display all checks by default,
but thanks to this, now the flags are parsed before the rules,
which makes possible not to search for the -rules flag bypassing
through findRulesFlag.

Now, to set all the rules in -allow-checks, you need to write '<all>'.