linttest: improved inline tests
Now each file is used only to check the checker
with the name of the file name equal, all other
errors will be ignored.

If the file has a name for which there is no
checker, an error will be thrown.

Fixed an error message if the want directive
is written for the line, but there is no actual
error for this line.