linter: refactored report functions
Now at the heart of all the functions for creating reports
is a function that creates a report at a given location.

Report and ReportPHPDoc are now wrappers that convert node
and PHPDocLocation to Location for ReportLocation.

ReportByLine has been removed as it is no longer used and
can be replaced with ReportLocation.

Fixed error, if the warning was given on the last line in
the file, then the context line was empty.

A new ir.Location structure has been added.
ir.Location stores the shifts relative to the current line.
A new structure is needed since position.Position stores
line shifts relative to the beginning of the file, and not
relative to the current line, which is very inconvenient for
using ReportLocation.