linter: added strict-mixed mode
Previously, by default, we would give an undefined method
or property error even if the type of the variable was mixed,
which led to the fact that poorly typed code gave a large
number of errors.

Now this behavior has been taken under the --strict-mixed flag,
without it in these cases an error will not be thrown.