linter, cmd: fixed incorrect work of quick fixes
Now the rules will not be applied to the vendor folder if it
got into the analysis, for example, if the analysis was started
on the root folder in which the vendor is located.

By default, now the vendor folder is also added to the --exclude.

For the propNullDefault checker, the addQuickFix function is used
instead of simple adding to the slice.