linter: added 'test-rules' command for testing dynamic rules
Each test file must have a name equivalent to the name of the
rule being tested. If the rule is in a namespace, then the
test must also be in that namespace.

Folders and files with rules are passed to the --rules flag,
one folder or file with tests is passed to the --tests flag.

By default, the --rules flag contains the current directory,
and the --tests flag contains the './tests' folder, which
corresponds to the location of the 'tests' folder in the
rules folder:

.
├── rules.php
│
└── tests
    └── testFile.php
    ...

Each test file is a code with special comments.
For those lines on which the linter gives warnings according to
the rules, a comment of the form "want `warning message`" is
written, which indicates to the linter that this warning is
expected here.

If there are more than one warnings on a line, then you can
add as many "and `warning 2`" constructions to the end as you
need.