phpgrep: fix functions alias matching
When matching things like `sizeof(...)` phpgrep would interpret
it as `count(...)`, since sizeof is just an alias.

It breaks when we actually need to get the `sizeof` identifier position
as `phpcore.ResolveAlias` returns a node without pos info.

This causes a runtime panic in phpgrep when such node is being rendered.

The current fix is to create a new node that has the resolved name
value and original source location info.